+++ packages/frontend/src/view/components/Button.tsx
@@ -11,7 +11,7 @@ type ButtonVariant = 'contained' | 'outlined'
 const mainClassNames =
   'py-2.5 text-sm font-semibold disabled:cursor-not-allowed'
 const classNameMap: Record<ButtonVariant, string> = {
-  contained: 'bg-brand w-32 rounded',
+  contained: 'bg-brand rounded',
   outlined:
     'bg-transparent border border-brand hover:bg-brand hover:bg-opacity-20 px-8 rounded-lg',
 }
