+++ packages/frontend/src/view/pages/user/components/UserQuickActionsTable.tsx
@@ -1,0 +1,84 @@
+import React from 'react'
+
+import { Asset, assetToInfo } from '../../../../utils/assets'
+import {
+  formatAmount,
+  formatWithDecimals,
+} from '../../../../utils/formatting/formatAmount'
+import { AssetWithLogo } from '../../../components/AssetWithLogo'
+import { Button } from '../../../components/Button'
+import { InlineEllipsis } from '../../../components/InlineEllipsis'
+import { UserOfferEntry } from './UserOffersTable'
+
+interface UserQuickActionsTableProps {
+  readonly withdrawableAssets: readonly WithdrawableAssetEntry[]
+  readonly offersToAccept: readonly UserOfferEntry[]
+  isMine?: boolean
+}
+
+export interface WithdrawableAssetEntry {
+  readonly asset: Asset
+  readonly amount: bigint
+}
+
+export function UserQuickActionsTable(props: UserQuickActionsTableProps) {
+  return (
+    <div className="mb-12 flex w-full flex-col rounded-lg border border-solid border-brand bg-gray-800 p-6">
+      <p className="text-sm font-semibold text-zinc-500">Withdrawable assets</p>
+      {props.withdrawableAssets.map((asset) => {
+        const assetInfo = assetToInfo(asset.asset)
+        return (
+          <div
+            className="mt-4 flex items-center justify-between"
+            key={assetInfo.symbol}
+          >
+            <AssetWithLogo assetInfo={assetInfo} type="regularSymbol" />
+            <p className="text-base text-zinc-500">
+              Finalize the withdrawal of{' '}
+              <strong className="text-white">
+                {formatAmount(asset.asset, asset.amount)}{' '}
+                <InlineEllipsis className="max-w-[80px]">
+                  {assetInfo.symbol}
+                </InlineEllipsis>
+              </strong>
+            </p>
+            <Button className="w-32">Withdraw now</Button>
+          </div>
+        )
+      })}
+      {props.isMine && (
+        <>
+          <p className="mt-6 text-sm font-semibold text-zinc-500">
+            Offers to accept
+          </p>
+          {props.offersToAccept.map((offer) => {
+            const assetInfo = assetToInfo(offer.asset)
+            const totalPrice = offer.amount * offer.price
+            return (
+              <div
+                className="mt-3 flex items-center justify-between"
+                key={offer.timestamp.toString()}
+              >
+                <AssetWithLogo assetInfo={assetInfo} type="regularSymbol" />
+                <p className="text-base text-zinc-500">
+                  Finalize the offer{' '}
+                  <strong className="text-white">
+                    {formatAmount(offer.asset, offer.amount)}{' '}
+                    <InlineEllipsis className="max-w-[80px]">
+                      {assetInfo.symbol}
+                    </InlineEllipsis>
+                  </strong>{' '}
+                  in exchange for{' '}
+                  <strong className="text-white">
+                    {formatWithDecimals(totalPrice, 6, { suffix: ' USDC' })}
+                  </strong>
+                </p>
+                <Button className="w-32">Accept & sell</Button>
+              </div>
+            )
+          })}
+        </>
+      )}
+    </div>
+  )
+}
